+++ processor/src/trace/trace_state.rs
@@ -29,6 +29,9 @@ use crate::{
 
 mod serde;
 
+#[cfg(test)]
+mod serialization_tests;
+
 // TRACE FRAGMENT CONTEXT
 // ================================================================================================